requirements.txt
alembic==1.10.3
anyio==3.6.2
Authlib==0.15.6
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.9.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.12.1
bcrypt==3.2.2
bidict==0.22.1
blinker==1.6.1
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
coverage==7.2.4
cryptography==3.4.8
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.1
eventlet==0.33.0
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-auth==0.1.1
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.6
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httpagentparser==1.9.5
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==2.10
importlib-metadata==6.3.0
importlib-resources==5.12.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
ldap3==2.9.1
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.2
orjson==3.8.10
packaging==23.1
paramiko==3.1.0
passlib==1.7.4
pgadmin4 @ https://ftp.postgresql.org/pub/pgadmin/pgadmin4/v6.15/pip/pgadmin4-6.15-py3-none-any.whl
Pillow==9.5.0
pluggy==1.0.0
portalocker==2.7.0
psutil==5.9.4
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyNaCl==1.5.0
pyotp==2.8.0
pypng==0.20220715.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2021.3
PyYAML==6.0
qrcode==7.4.2
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
speaklater3==1.4
SQLAlchemy==1.4.47
SQLAlchemy-serializer==1.4.1
SQLAlchemy-Utils==0.41.0
sqlparse==0.4.3
sshtunnel==0.4.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
ua-parser==0.16.1
ujson==5.7.0
urllib3==1.26.15
user-agents==2.2.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.15.0
